moved Knex file to root directory
